+++ automatic/cygwin/tools/chocolateyInstall.ps1
@@ -1,12 +1,13 @@
 ﻿# https://cygwin.com/faq/faq.html#faq.setup.cli
 
 $ErrorActionPreference = 'Stop'
+$toolsPath = "$(Split-Path -parent $MyInvocation.MyCommand.Definition)"
 
 $pp = Get-PackageParameters
 
 $cygwin_root = (Get-ItemProperty 'HKLM:\SOFTWARE\Cygwin\setup' -ea 0).rootdir
 if (!$cygwin_root) {
-    $cygwin_root = if ($pp.InstallDir) { $pp.InstallDir } else { (Get-BinRoot) + '\cygwin' }
+    $cygwin_root = if ($pp.InstallDir) { $pp.InstallDir } else { (Get-ToolsLocation) + '\cygwin' }
 } else { Write-Host 'Existing installation detected, ignoring InstallDir argument' }
 
 if (!$pp.Proxy) {
@@ -35,22 +36,18 @@ $silentArgs = @(
 )
 
 $packageArgs = @{
-  packageName            = 'Cygwin'
-  fileType               = 'exe'
-  url                    = 'https://cygwin.com/setup-x86.exe'
-  url64bit               = 'https://cygwin.com/setup-x86_64.exe'
-  checksum               = '406bfa31d0c724de90aced5edddaf07f3eac46220c9cc9998ae2428d13989c28'
-  checksum64             = 'b67afe083b2547acee0857b45ff9784f873d1fb226a0148c1364f255180fc282'
-  checksumType           = 'sha256'
-  checksumType64         = 'sha256'
-  silentArgs             = $silentArgs
-  validExitCodes         = @(0)
+  packageName    = $env:ChocolateyPackageName
+  fileType       = 'exe'
+  file           = "$toolsPath\"
+  file64         = "$toolsPath\"
+  softwareName   = 'Cygwin*'
+  silentArgs     = $silentArgs
+  validExitCodes = @(0)
 }
-Install-ChocolateyPackage @packageArgs
+Install-ChocolateyInstallPackage @packageArgs
+
 Install-BinFile -Name "Cygwin" -Path "$cygwin_root\Cygwin.bat"
 
 Write-Host "Copying cygwin package manager (setup) to $cygwin_root"
-$setup_path = if (Get-ProcessorBits 64) { $packageArgs.url64bit } else { $packageArgs.url }
-$setup_path = "{0}\{1}" -f (Get-PackageCacheLocation), ($setup_path -split '/' | select -Last 1)
-if (!$setup_path)  { Write-Warning "Can't find setup path"; return }
-cp $setup_path $cygwin_root\cygwinsetup.exe
+$setup_path = if (Get-ProcessorBits 64 -and $env:ChocolateyForceX86 -ne $true) { $packageArgs.file } else { $packageArgs.file64 }
+mv $setup_path $cygwin_root\cygwinsetup.exe -Force