+++ automatic/cygwin/tools/chocolateyUninstall.ps1
@@ -1,4 +1,4 @@
-$ErrorActionPreference = 'Stop'
+﻿$ErrorActionPreference = 'Stop'
 
 $cygwin_root = (Get-ItemProperty 'HKLM:\SOFTWARE\Cygwin\setup' -ea 0).rootdir
 rm 'HKLM:\SOFTWARE\Cygwin' -Recurse -ErrorAction SilentlyContinue